+++ PathFinding/PathFinding/Edge.h
@@ -11,15 +11,15 @@ class Edge {
 private:
 	int m_node1;
 	int m_node2;
-	int m_weight;
+	float m_weight;
 public:
 	Edge();
-	Edge(int, int, int);
+	Edge(int, int, float);
 
 
-	int get_weight();
+	float get_weight();
 	void extract_nodes(int&, int&);
-	void set_weight(int);
+	void set_weight(float);
 	void set_nodes(int, int);
 	bool get_pair_weight(int, int&, float&);
 	friend ostream& operator << (ostream&, const Edge&);
